pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "DILIGENT"
version = "2.1"
requires-python = ">=3.14"

dependencies = [
    "fastapi==0.128.0",
    "uvicorn[standard]>=0.40.0",
    "numpy==2.4.1",
    "pandas==3.0.0",
    "tqdm==4.67.1",
    "SQLAlchemy==2.0.46",
    "psycopg[binary]==3.3.1",
    "httpx==0.27.2",    
    "pydantic-settings==2.11.0",
    "python-dotenv==1.2.1",
    "cryptography==46.0.3",
    "rapidfuzz==3.9.4",   
    "pdfminer.six==20250506",
    "pypdf==6.1.1",    
    "langchain==1.2.0",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.1.0",
    "lancedb==0.26.0",
    "openpyxl==3.1.5",    
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-playwright>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["DILIGENT"]




